Moved DiscontinuityEmitter impl. duplication from contributors to RoomLifeCycleManager ContributesToRoomLifecycleImpl abstract class
