+++ chat-android/src/main/java/com/ably/chat/Typing.kt
@@ -2,7 +2,6 @@
 
 package com.ably.chat
 
-import io.ably.lib.types.ErrorInfo
 import io.ably.lib.realtime.Channel as AblyRealtimeChannel
 
 /**
@@ -79,7 +78,7 @@ data class TypingEvent(val currentlyTyping: Set<String>)
 internal class DefaultTyping(
     roomId: String,
     private val realtimeClient: RealtimeClient,
-) : Typing, ContributesToRoomLifecycle, ResolvedContributor {
+) : Typing, ContributesToRoomLifecycleImpl(), ResolvedContributor {
 
     private val typingIndicatorsChannelName = "$roomId::\$chat::\$typingIndicators"
 
@@ -90,8 +89,6 @@ internal class DefaultTyping(
     override val attachmentErrorCode: ErrorCodes = ErrorCodes.TypingAttachmentFailed
 
     override val detachmentErrorCode: ErrorCodes = ErrorCodes.TypingDetachmentFailed
-
-    private val discontinuityEmitter = DiscontinuityEmitter()
 
     override fun subscribe(listener: Typing.Listener): Subscription {
         TODO("Not yet implemented")
@@ -107,16 +104,5 @@ internal class DefaultTyping(
 
     override suspend fun stop() {
         TODO("Not yet implemented")
-    }
-
-    override fun onDiscontinuity(listener: EmitsDiscontinuities.Listener): Subscription {
-        discontinuityEmitter.on(listener)
-        return Subscription {
-            discontinuityEmitter.off(listener)
-        }
-    }
-
-    override fun discontinuityDetected(reason: ErrorInfo?) {
-        discontinuityEmitter.emit("discontinuity", reason)
     }
 }
